+++ app/dashboard/income/_actions/deleteIncome.ts
@@ -2,7 +2,6 @@
 
 import prisma from "@/prisma/prisma";
 import { auth } from "@clerk/nextjs/server";
-import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 export async function DeleteIncome(id: number) {
@@ -16,13 +15,63 @@ export async function DeleteIncome(id: number) {
     where: { clerkId: userId },
   });
 
-  await prisma.budget.delete({
+  if (!existingUser) {
+    throw new Error("User not found.");
+  }
+
+  const deletedBudget = await prisma.budget.delete({
     where: {
-      clerkId: userId,
-      userId: existingUser?.id,
       id,
+      clerkId: userId,
+      userId: existingUser.id,
     },
   });
 
-  revalidatePath("/dashboard/income");
+  if (!deletedBudget) {
+    throw new Error("Failed to delete the budget entry.");
+  }
+
+  const [budget, budgetRules, totalFixedExpenses] = await prisma.$transaction([
+    prisma.budget.aggregate({
+      where: { clerkId: userId, userId: existingUser.id },
+      _sum: { amount: true },
+    }),
+    prisma.budgetRule.findFirst({
+      where: { clerkId: userId, userId: existingUser.id },
+    }),
+    prisma.fixedExpense.aggregate({
+      where: { clerkId: userId, userId: existingUser.id },
+      _sum: { budgetAmount: true },
+    }),
+  ]);
+
+  const totalBudget = budget._sum.amount || 0;
+  const totalFixed = totalFixedExpenses._sum.budgetAmount || 0;
+
+  if (totalBudget > 0) {
+    const needsPercentage = (totalFixed / totalBudget) * 100;
+
+    const updatedBudgetRule = await prisma.budgetRule.upsert({
+      where: { id: budgetRules?.id || 0 },
+      update: {
+        actualNeedsPercentage: needsPercentage,
+        actualSavingsPercentage: 0,
+        actualWantsPercentage: 0,
+      },
+      create: {
+        needsPercentage: 50,
+        savingsPercentage: 30,
+        wantsPercentage: 20,
+        actualNeedsPercentage: needsPercentage,
+        actualSavingsPercentage: 0,
+        actualWantsPercentage: 0,
+        userId: existingUser.id,
+        clerkId: userId,
+      },
+    });
+
+    return updatedBudgetRule;
+  }
+
+  throw new Error("Budget calculations failed.");
 }
